# lib/api-zod/src/generated/types/createQaRunRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQaRunRequest {
  appUrl: string;
  /** @minLength 10 */
  appDescription: string;
}