# lib/api-zod/src/generated/types/qaIssue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QaIssueSeverity } from "./qaIssueSeverity";

export interface QaIssue {
  title: string;
  description: string;
  severity: QaIssueSeverity;
  possibleCause: string;
  suggestedFix: string;
}