# lib/api-zod/src/generated/types/qaIssueSeverity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type QaIssueSeverity =
  (typeof QaIssueSeverity)[keyof typeof QaIssueSeverity];

export const QaIssueSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;