# lib/api-zod/src/generated/types/qaReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QaIssue } from "./qaIssue";

export interface QaReport {
  summary: string;
  issues: QaIssue[];
  /**
   * @minimum 0
   * @maximum 100
   */
  overallScore: number;
  recommendations: string[];
  /** @nullable */
  screenshotBase64?: string | null;
}