# lib/api-zod/src/generated/types/qaRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QaRunStatus } from "./qaRunStatus";

export interface QaRun {
  id: string;
  userId: string;
  appUrl: string;
  appDescription: string;
  status: QaRunStatus;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}