# lib/api-zod/src/generated/types/qaRunStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type QaRunStatus = (typeof QaRunStatus)[keyof typeof QaRunStatus];

export const QaRunStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;