# lib/api-zod/src/generated/types/qaRunWithReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QaReport } from "./qaReport";
import type { QaRun } from "./qaRun";

export type QaRunWithReport = QaRun & {
  report?: QaReport | null;
};